refactor: use component for search page
which eases up adding a new locale a LOT
